Version 1.0.35
Important: Some internal changes were made to the Send command. If any of your scripts rely on the subtle differences between the left, right, and neutral modifiers keys (Control, Alt, Shift, and Win), it is recommended that they be retested prior to adopting this version.

Fixed the Send command's inability to send AltGr characters (such as "\" and "{" on the German keyboard) to apps that don't allow Ctrl+Alt as a substitute for AltGr. [thanks xrat]

Improved functions to support optional parameters.

GUI:

Fixed Tab controls' g-label to launch when the page is changed via keyboard navigation. Also, controls added to a Tab control while the window is visible are drawn correctly.

Improved Edit controls to support g-labels, which notify the script when an Edit's contents have changed.

Added control types DateTime and MonthCal, which allow a date and/or time to be entered in various formats.

Added control type UpDown, which is a pair of arrow buttons that can be clicked to increase or decrease a value.