Version 1.0.36
Fixed double syntax error for a bad function call in a block.

Improved "Process, Exist" to fetch the script's own PID when no other parameters are given. [thanks Junyx]

CHANGES FOR GUI

Added GuiContextMenu, which facilitates the display of custom Gui context menus.

Added Gui +Delimiter`n as a means to change the separator between fields.

Added A_EventInfo for use by OnClipboardChange, ListBox, GuiSize, and GuiDropFiles.

Added A_GuiEvent as a synonym for A_GuiControlEvent.

Added +/-Redraw as a means to accelerate the filling of a ListBox.

Added new control type ListView, which is a tabular view of rows and columns. It features built-in sorting and support for displaying icons.