Version 1.0.36.01
Fixed FileCopy and FileMove to avoid deleting a file that is copied/moved onto itself. In such a case, FileCopy still counts it as a failure but FileMove now counts it as a success. [thanks mario_a]

Fixed continuation sections so that accents and commas after the word Join don't change the behavior of accents and commas in the section. [thanks Nemroth]

Added "WinSet Top", which attempts to bring a window to the top of the stack without activating it. [thanks catweazle]