Version 1.0.36.02
Fixed hotstrings to work when the script is displaying an InputBox.

Fixed tilde key-up hotkeys such as "~5 up::" to avoid double-character output.

Disallowed linefeeds (`n) and other invisible characters in variable names.

Increased the stack limit from 1 to 4 MB, which allows functions and subroutines to recurse more deeply. Note: This does not increase the memory used by a typical script. [thanks Rubens]

Doubled the speed of accessing empty variables (which are slow due to the need to check if they're environment variables).

Improved all windowing commands to support multiple criteria (such as title+class) to narrow down the search.

Added built-in variable A_LoopFileExt, which contains the loop file's extension.