Version 1.0.36.07
Fixed tilde key-up hotkeys for Win/Alt such as "~LWin up::".

Fixed bitwise-not of a floating point value such as ~1.0.

Improved VarSetCapacity with the ability to zero-fill a variable's memory.

Improved DllCall by adding Gdi32.dll to the set of default DLLs.

Added operators & and * for use with memory addresses and binary data.

Added documentation to DllCall demonstrating how to use structures.