Version 1.0.37.05
Fixed hotkeys such as ^!a not to fire twice when "a & b::" is a hotkey. Also fixed hotkeys such as ^!a to work without a $ when "a" is a prefix key such as "a & b::". [thanks t0nyk0]

Fixed hotkeys "LButton::" and "RButton::" not to disrupt Edit control context menus. [thanks Tekl]

Fixed "GuiControl ChooseString" not to ignore the first two items in each ListBox/ComboBox/DDL. [thanks Tekl]