Version 1.0.38.06
Some optimizations were made to the keyboard and mouse hooks. If you have any mission-critical scripts that rely on them, it is recommended that they be re-tested or that you wait two weeks for any bugs to get fixed.

Fixed continuation lines that start with "AND" and "OR" to work even when the next character is in the set "()!~*&-+". [thanks Decarlo110]

Fixed hotkeys/labels such as "(::" and "(MyLabel):" not to be misinterpreted as continuation sections. [thanks whismerhill]