Version 1.0.39
The last change below may impact scripts that use the keyboard or mouse hook. If you have any mission-critical scripts that rely on them, it is recommended that they be re-tested or that you wait two weeks for any bugs to get fixed.

Fixed inability to have lines consisting entirely of spaces and tabs in continuation sections. [thanks JSLover]

Fixed hotkeys such as "LButton & LCtrl" to auto-install the mouse hook. [thanks JSLover]

Added a dedicated thread to service the keyboard and mouse hooks. This eliminates mouse/keyboard lag during long operations such as UrlDownloadToFile. It also prevents keys from getting "stuck down". The new thread consumes only 36 KB of memory and it exists only in scripts that use the hooks.