Version 1.0.40
Fixed the v1.0.39 mouse and keyboard hooks so that failure to activate them behaves correctly.

Changed the Send command for Russian and other keyboard layouts to be able to produce more ASCII characters (such as the letters A to Z). This does not affect most Western European and English layouts.

Changed hotkeys that are a subset of a wildcard hotkey to take precedence. For example, if *x and ^x are both hotkeys, pressing ^x will now trigger ^x rather than *x.

Improved syntax so that double colons do not need to be escaped.

Improved Send and ControlSend with option {Blind}, which leaves Ctrl/Alt/Shift/Win pressed down if they started that way.

Added a new remapping method that is more simple and powerful than the old methods. For example:

a::b  ; Make the "a" key become "b".
Capslock::Ctrl  ; Make Capslock become a Control key.
XButton1::LButton  ; Make the fourth mouse button behave like the left mouse button.
RCtrl::RWin  ; Make the right Control key become the right Windows key.