Version 1.0.42.03
Fixed crash of WinActivate in certain rare circumstances. [thanks twhyman]

Changed hotstrings to require the mouse hook by default (see next item). This can be prevented via #Hotstring NoMouse.

Improved hotstrings to take into account mouse clicks. This allows a hotstring to fire when the user's click focuses a new control or moves the caret.

Improved the Random command with a means to set a new seed. [thanks Laszlo]

Improved #ClipboardTimeout to reattempt data reading when the first attempt fails. Previously, only the opening of the clipboard was reattempted.

Added built-in variable A_LastError for DllCall and Run/RunWait. It contains the result from the OS's GetLastError().