Version 1.0.44.03
Fixed FileExist() to report "X" rather than "" for files whose attributes are all undefined. [thanks Peter]

Fixed Tab controls so that the first Text control added to a new page is auto-positioned correctly even if the previously added control was a Text control on some other page.

Fixed ImageSearch so that a transparent color like TransBlack is as effective in variation mode as non-variation mode. [thanks Troz]

Fixed GetKeyState and KeyWait to accept characters that don't exist in the current keyboard layout (in case the keyboard layout changes after launching the script).

Fixed hotkeys that don't exist in the current keyboard layout (such as ^!&auml;:: in English) to display a warning and end the auto-execute section. [thanks Androgen Belkin]

Changed and improved ControlSend, Send, Hotstrings, Input, and AltGr handling to use the target window's language rather than the script's. If you use only one keyboard layout on your system, this should not affect anything. [thanks Androgen Belkin]

Improved VarSetCapacity() to interpret a capacity of -1 as "update this variable's internally-stored length". This is useful in cases where a variable has been altered indirectly, such as by passing its address via DllCall().